updated tags mutations and errors catching

+++ app/graphql/mutations/tags/delete_tag.rb
@@ -6,7 +6,10 @@ module Mutations
       field :message, String, null: false
 
       def resolve(id:)
-        tag = Tag.find(id)
+        tag = Tag.find_by(id: id)
+
+        raise GraphQL::ExecutionError, "Tag not found" unless tag.present?
+        raise GraphQL::ExecutionError, "You cannot delete this tag" unless tag.user_id == current_user.id
 
         tag.destroy
